Recruitment_Portal/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
WTForms==3.0.1
Pillow==11.2.1
markupsafe==2.1.3

# Dépendances pour SQL Server
pyodbc==5.2.0
SQLAlchemy-Utils==0.41.1
